student_tracker/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-filter==25.1
django-heroku==0.3.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
whitenoise==6.9.0
XlsxWriter==3.2.2
